Removing suppressed columns get function
