+++ src/csvcubed/cli/inspect/metadataprinter.py
@@ -145,8 +145,6 @@ class MetadataPrinter:
         """
         assert isinstance(self.state, DataCubeInspector)  # Make pyright happier
 
-        csvw_inspector = self.state.csvw_inspector
-
         self.result_qube_components = self.state.get_dsd_qube_components_for_csv(
             self.primary_csv_url
         )
@@ -160,9 +158,15 @@ class MetadataPrinter:
             )
         )
 
-        self.primary_csv_suppressed_columns = self.state.get_suppressed_columns_for_csv(
-            self.primary_csv_url
-        )
+        self.primary_csv_suppressed_columns = [
+            column_definition.title
+            for column_definition in self.primary_csv_column_definitions
+            if column_definition.suppress_output and column_definition.title is not None
+        ]
+
+        # self.primary_csv_suppressed_columns = self.state.get_suppressed_columns_for_csv(
+        #     self.primary_csv_url
+        # )
 
         self.result_primary_csv_code_lists = self.state.get_code_lists_and_cols(
             self.primary_csv_url
